+++ examples/ovault-evm/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @layerzerolabs/ovault-evm-example
 
+## 0.0.5
+
+### Patch Changes
+
+- 79ee352: Fix send task to use deploymentEid instead of just eid
+
+## 0.0.4
+
+### Patch Changes
+
+- 44cfbbb: Chore: improve the setup process for share only deployers.
+
+## 0.0.3
+
+### Patch Changes
+
+- e2eece6: Update RPCs to testnet and add feature flag to README
+
 ## 0.0.2
 
 ### Patch Changes