+++ packages/ovault-evm/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/ovault-evm
 
+## 0.0.4
+
+### Patch Changes
+
+- b263ec2: Export Types for OVault Typescript SDK
+
+## 0.0.3
+
+### Patch Changes
+
+- e16fa4d: update sdk for quoteSend function interface on the composer
+
 ## 0.0.2
 
 ### Patch Changes